requirements.txt
PyYAML>=6.0
asyncio-mqtt>=0.11.0
colorama>=0.4.6
cryptography>=41.0.0
geoip2>=4.6.0
psutil>=5.9.0
scapy>=2.5.0
twisted>=22.10.0
pymongo>=4.3.0
pymysql>=1.0.0
psycopg2-binary>=2.9.0
redis>=4.5.0
elasticsearch>=8.8.0
pymemcache>=4.0.0
paramiko>=3.1.0
flask>=2.3.0
flask-cors>=4.0.0
requests>=2.31.0
python-socketio>=5.8.0
schedule>=1.2.0
tabulate>=0.9.0
prometheus-client>=0.17.0
